Make plugins loading from initial state
- Make all plugins loaded at runtime
- scripLoader is now promise based
- initial state now also contains plugins
- Plugin loading status in state
- TODO: make loading plugins from UI work again
- TODO: make history length configurable
