--- .babelrc
@@ -1,10 +1,0 @@
-{
-  "presets": [
-    "react",
-    "es2015"
-  ],
-  "plugins": [
-    "transform-object-rest-spread",
-    "transform-flow-comments"
-  ]
-}
