+++ package.json
@@ -10,9 +10,10 @@
     "start": "node bin/iot-dashboard.js",
     "test": "gulp test",
     "clean": "gulp clean && npm prune",
-    "build": "gulp build",
+    "build": "npm run compile:plugins && gulp build",
     "release": "npm run clean && npm run build --production && npm publish",
     "compile": "gulp compile",
+    "compile:plugins": "babel ./plugins --out-dir ./dist/plugins",
     "dev": "gulp dev",
     "watch": "gulp watch",
     "lint": "gulp lint",
@@ -39,6 +40,15 @@
     "sensor",
     "monitoring"
   ],
+  "babel": {
+    "presets": [
+      "react",
+      "es2015"
+    ],
+    "plugins": [
+      "transform-flow-comments"
+    ]
+  },
   "dependencies": {
     "connect": "^3.4.1",
     "serve-static": "^1.11.1"
@@ -49,7 +59,6 @@
     "babel-eslint": "6.0.4",
     "babel-loader": "6.2.4",
     "babel-plugin-transform-flow-comments": "6.8.0",
-    "babel-plugin-transform-object-rest-spread": "6.6.5",
     "babel-preset-es2015": "6.6.0",
     "babel-preset-react": "6.5.0",
     "chai": "3.5.0",
