+++ src/actionNames.ts
@@ -38,6 +38,7 @@ export const ADD_WIDGET_PLUGIN = "ADD_WIDGET_PLUGIN";
 export const ADD_DATASOURCE_PLUGIN = "ADD_DATASOURCE_PLUGIN";
 export const DELETE_WIDGET_PLUGIN = "DELETE_WIDGET_PLUGIN";
 export const DELETE_DATASOURCE_PLUGIN = "DELETE_DATASOURCE_PLUGIN";
+export const PLUGIN_IS_LOADING = "PLUGIN_IS_LOADING";
 
 // Modal
 export const SHOW_MODAL = "SHOW_MODAL";
