+++ src/appState.ts
@@ -22,6 +22,7 @@ export interface State {
     widgets: Widgets.IWidgetsState
     datasources: Datasource.IDatasourcesState
     datasourcePlugins: DatasourcePlugins.IDatasourcePluginsState
+    widgetPlugins: any // TODO: type WidgetPlugins (and migrate to ts)
 }
 
 export interface ITypeInfo {
@@ -35,10 +36,10 @@ export interface ITypeInfo {
 export interface ISetting {
     id: string // Technical id, used to receive the value later
     name: string // User friendly string to describe the value
-    description: string // User friendly description with detail information about the value
-    defaultValue: string // The default value that is preset when a new Plugin is configured, currently must be a string
-    required: boolean // true when the setting is required
     type: string // Defines how the setting is rendered, validated and interpreted
+    description?: string // User friendly description with detail information about the value
+    defaultValue?: any // The default value that is preset when a new Plugin is configured, currently must be a string
+    required?: boolean // true when the setting is required
 }
 
 export default State;
