+++ src/browser-tests.ts
@@ -14,7 +14,6 @@ import './datasource/datasource.test.ts'
 import './datasource/datasourcePluginFactory.test.ts'
 import './datasource/datasourcePluginRegistry.test.ts'
 import './datasource/datasourcePlugins.test.ts'
-import './datasource/plugins/randomDatasource.test.js'
 import './pluginApi/uri.test.js'
 import './serverRenderer.test.ts'
 import './util/collection.test.js'
