+++ src/config.json
@@ -1,6 +1,6 @@
 {
   "version": "0.0.4",
-  "revision": "77743a7a751e8fe58e72803ba352d0d3a26b7f43",
-  "revisionShort": "77743a7",
+  "revision": "d055da2f41adca6c379e554d1a4a0569c5cec964",
+  "revisionShort": "d055da2",
   "branch": "getting-started"
 }
