+++ src/typings/loadjs/index.d.ts
@@ -1,10 +1,7 @@
-
-
-
-
 declare module "loadjs" {
     interface Options {
         success: () => void;
+        error: (error: Error) => void;
     }
     function loadjs(paths: Array<string>, options: Options): void;
